refactor:カテゴリー登録フォームをコンポーネント化 #68
